backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ffmpeg-python==0.2.0
gunicorn==21.2.0
python-multipart==0.0.6
